fastapi/fastapi-ai/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pyyaml==6.0.2
nacos-sdk-python==2.0.8
requests==2.32.4
sqlmodel==0.0.24
pymysql==1.1.1
jieba==0.42.1
cozepy==0.19.0